Add files via upload
The emergence of analytics poses a world of almost infinite opportunities for sectors such as insurance, where intelligence has long been focused on enterprises. Because of cost considerations, the market at large has had a slow acceptance of fresh Big Data analytics, and regulation could be the future's controlling pressure. 
In the past, identification of fraud was restricted to agents of claims who had to rely on little facts and a great amount of intuition. New data analysis has introduced tools to make fraud review and detection possible in other areas such as underwriting, policy renewals, and periodic checks that ﬁt right in with modeling. 
This project aims at insurance fraud detection based on a stacking classifier and evaluated on the AUC curve.  